Applied a patch from Matthias Vogelgesang (KIT)
git-svn-id: svn+ssh://svn.code.sf.net/p/tango-cs/code/share/yat/trunk@21189 4e9c00fd-8f2e-0410-aa12-93ce3db5e235

+++ src/file/PosixFileImpl.cpp
@@ -489,7 +489,7 @@ void FileName::copy(const String &strDst, bool bKeepMetaData) throw( Exception )
     {
       size_t lToRead = 0;
 
-      if( llSize > s_copy_bloc_size )
+      if( (size_t)llSize > s_copy_bloc_size )
         lToRead = s_copy_bloc_size;
       else
         lToRead = (size_t)llSize;
